# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Book Library API",
    "description": "Full CRUD API for managing Books and Authors - CSE-341 Project 2 (with Google OAuth)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://cse-341-project2-ea66.onrender.com",
      "description": "Production (Render)"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local development"
    }
  ],
  "paths": {
    "/api/auth/google": {
      "get": {
        "summary": "Login with Google OAuth",
        "description": "Redirects user to Google for authentication",
        "responses": {
          "302": {
            "description": "Redirect to Google login"
          }
        }
      }
    },
    "/api/auth/google/callback": {
      "get": {
        "summary": "Google OAuth Callback",
        "description": "Handles the callback from Google after authentication",
        "responses": {
          "302": {
            "description": "Redirect to Swagger UI after successful login"
          }
        }
      }
    },
    "/api/auth/current-user": {
      "get": {
        "summary": "Get current logged-in user",
        "responses": {
          "200": {
            "description": "Returns current user information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "displayName": { "type": "string" },
                    "email": { "type": "string" },
                    "photo": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Not authenticated" }
        }
      }
    },
    "/api/auth/logout": {
      "get": {
        "summary": "Logout user",
        "responses": {
          "200": { "description": "Logged out successfully" }
        }
      }
    },

    "/api/books": {
      "get": {
        "summary": "Get all books",
        "description": "Returns a list of all books sorted by addedDate (newest first)",
        "responses": {
          "200": {
            "description": "Array of books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Book" }
                }
              }
            }
          },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "summary": "Create a new book",
        "description": "Creates a new book with validation (Requires Google OAuth authentication)",
        "security": [{ "sessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookInput" },
              "example": {
                "title": "Atomic Habits",
                "author": "James Clear",
                "isbn": "9780735211292",
                "genre": "Self-Help",
                "publicationYear": 2018,
                "pageCount": 320,
                "description": "A guide on how small habits can lead to big improvements in life.",
                "coverImageUrl": "https://images-na.ssl-images-amazon.com/images/I/91bYsX41DVL.jpg",
                "rating": 4.8,
                "isRead": true,
                "userNotes": "Helped me build better daily routines."
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Book" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized - Please log in with Google" },
          "409": { "description": "Duplicate ISBN" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/books/{id}": {
      "get": {
        "summary": "Get book by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" },
            "description": "MongoDB ObjectId (24 hexadecimal characters)"
          }
        ],
        "responses": {
          "200": {
            "description": "Book object",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Book" }
              }
            }
          },
          "400": { "description": "Invalid ID format" },
          "404": { "description": "Book not found" },
          "500": { "description": "Server error" }
        }
      },
      "put": {
        "summary": "Update book",
        "description": "Update an existing book (Requires Google OAuth authentication)",
        "security": [{ "sessionAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated book",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Book" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Book not found" },
          "500": { "description": "Server error" }
        }
      },
      "delete": {
        "summary": "Delete book",
        "description": "Delete a book (Requires Google OAuth authentication)",
        "security": [{ "sessionAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted successfully" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Book not found" },
          "500": { "description": "Server error" }
        }
      }
    },

    "/api/authors": {
      "get": {
        "summary": "Get all authors",
        "responses": {
          "200": {
            "description": "Array of authors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Author" }
                }
              }
            }
          },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "summary": "Create a new author",
        "description": "Create a new author (Requires Google OAuth authentication)",
        "security": [{ "sessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthorInput" },
              "example": {
                "name": "James Clear",
                "bio": "Author of Atomic Habits",
                "nationality": "American",
                "birthYear": 1985,
                "website": "https://jamesclear.com"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Author created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Author" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/api/authors/{id}": {
      "get": {
        "summary": "Get author by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Author object",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Author" }
              }
            }
          },
          "404": { "description": "Not found" },
          "500": { "description": "Server error" }
        }
      },
      "put": {
        "summary": "Update author",
        "description": "Update an existing author (Requires Google OAuth authentication)",
        "security": [{ "sessionAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthorInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated author",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Author" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" },
          "500": { "description": "Server error" }
        }
      },
      "delete": {
        "summary": "Delete author",
        "description": "Delete an author (Requires Google OAuth authentication)",
        "security": [{ "sessionAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted successfully" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" },
          "500": { "description": "Server error" }
        }
      }
    }
  },

  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie automatically set after successful Google OAuth login"
      }
    },
    "schemas": {
      "Book": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "author": { "type": "string" },
          "isbn": { "type": "string" },
          "genre": {
            "type": "string",
            "enum": ["Fiction", "Non-Fiction", "Sci-Fi", "Fantasy", "Mystery", "Biography", "Self-Help", "Psychology", "Business", "History", "Other"]
          },
          "publicationYear": { "type": "integer", "minimum": 1000 },
          "pageCount": { "type": "integer", "minimum": 1 },
          "description": { "type": "string", "maxLength": 1000 },
          "coverImageUrl": { "type": "string" },
          "rating": { "type": "number", "minimum": 0, "maximum": 5 },
          "isRead": { "type": "boolean" },
          "userNotes": { "type": "string" },
          "addedDate": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "BookInput": {
        "type": "object",
        "required": ["title", "author"],
        "properties": {
          "title": { "type": "string" },
          "author": { "type": "string" },
          "isbn": { "type": "string" },
          "genre": {
            "type": "string",
            "enum": ["Fiction", "Non-Fiction", "Sci-Fi", "Fantasy", "Mystery", "Biography", "Self-Help", "Psychology", "Business", "History", "Other"]
          },
          "publicationYear": { "type": "integer", "minimum": 1000 },
          "pageCount": { "type": "integer", "minimum": 1 },
          "description": { "type": "string", "maxLength": 1000 },
          "coverImageUrl": { "type": "string" },
          "rating": { "type": "number", "minimum": 0, "maximum": 5 },
          "isRead": { "type": "boolean" },
          "userNotes": { "type": "string" }
        }
      },
      "Author": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "bio": { "type": "string", "maxLength": 1000 },
          "nationality": { "type": "string" },
          "birthYear": { "type": "integer", "minimum": 1500 },
          "website": { "type": "string" },
          "photoUrl": { "type": "string" },
          "addedDate": { "type": "string", "format": "date-time" }
        }
      },
      "AuthorInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "bio": { "type": "string", "maxLength": 1000 },
          "nationality": { "type": "string" },
          "birthYear": { "type": "integer", "minimum": 1500 },
          "website": { "type": "string" },
          "photoUrl": { "type": "string" }
        }
      }
    }
  }
}